add check files exist

+++ webapp/deploy.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/env bash
 #upload files
-npm run build &&
-aws s3 rm s3://futboldesa --recursive &&
-aws s3 cp ./dist s3://futboldesa --recursive --acl public-read
+npm run build
+if [ -e  ./dist/index.html ] # check index exists
+then
+  aws s3 rm s3://futboldesa --recursive
+  aws s3 cp ./dist s3://futboldesa --recursive --acl public-read
+fi
